11/11/2015
Performed autocorrelation testing on the _sumrec Stratum files
Performed MannKendall tests on the Stratum files
Tested residuals for homskedasticity. As discovered with the Zone
_sumrec files, natural log transformation was best for the _sumrec
Stratum CPUE. I added 0.01 to all series to deal with presence of zeros
when transforming.
Also set up correlation scheme on line 447 to prepare Stratum
matrices..